requirements.txt
openai>=1.58.0
autogen>=0.2.0
flaml[automl]>=1.2.8
streamlit==1.29.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
pillow>=8.3.2,<10.1.0
numpy==1.24.3
decorator==4.4.2
proglog==0.1.10
tqdm==4.66.1 